+++ lib/src/tab/scripts/powershell.dart
@@ -11,7 +11,11 @@
 import '../flags.dart';
 import '_utils.dart';
 
-String powershellScript(String name, String exec) {
+String powershellScript(
+  String name,
+  String exec, {
+  bool enableDefaultCompletion = false,
+}) {
   validateShellCommandName(name);
   final escapedName = nameForVar(name);
   final execWords = powershellWords(exec);